address review comments; pull out message into separate component
